+++ controllers/service/controller.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/carv-ics-forth/frisbee/api/v1alpha1"
 	"github.com/carv-ics-forth/frisbee/controllers/common"
+	"github.com/carv-ics-forth/frisbee/controllers/common/watchers"
 	"github.com/carv-ics-forth/frisbee/pkg/lifecycle"
 	"github.com/go-logr/logr"
 	cmap "github.com/orcaman/concurrent-map"
@@ -32,7 +33,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -237,7 +237,7 @@ func (r *Controller) Finalize(obj client.Object) error {
 */
 
 func NewController(mgr ctrl.Manager, logger logr.Logger) error {
-	controller := &Controller{
+	reconciler := &Controller{
 		Manager:           mgr,
 		Logger:            logger.WithName("service"),
 		gvk:               v1alpha1.GroupVersion.WithKind("Service"),
@@ -245,9 +245,14 @@ func NewController(mgr ctrl.Manager, logger logr.Logger) error {
 		regionAnnotations: cmap.New(),
 	}
 
+	var (
+		service v1alpha1.Service
+		pod     corev1.Pod
+	)
+
 	return ctrl.NewControllerManagedBy(mgr).
+		For(&service).
 		Named("service").
-		For(&v1alpha1.Service{}).
-		Owns(&corev1.Pod{}, builder.WithPredicates(controller.Watchers())).
-		Complete(controller)
+		Owns(&pod, watchers.WatchWithPointAnnotation(reconciler, reconciler.gvk)).
+		Complete(reconciler)
 }